App restructured into /views folder with App.tsx routing
