+++ frontend/src/assets/PulseNotesLogo.png
