+++ frontend/src/views/TranscriptionPage.css
